backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
google-generativeai
pydantic-settings
alembic
python-multipart